Ensure correct value type in bucket (#162)
* Ensure correct value type in bucket

* Ensure correct value type in bucket

* Ensure correct value type in bucket

* Ensure correct value type in bucket

---------

+++ src/Analyzer/Processor/ParentPages.php
@@ -50,7 +50,7 @@ class ParentPages extends ProcessorBase {
 		$pageId = $this->xmlHelper->getIDNodeValue( $objectNode );
 		$parentPageId = $this->xmlHelper->getPropertyValue( 'parent', $objectNode );
 		if ( $parentPageId !== null ) {
-			$this->data['analyze-page-id-to-parent-page-id-map'][$pageId] = trim( $parentPageId );
+			$this->data['analyze-page-id-to-parent-page-id-map'][$pageId] = (int)trim( $parentPageId );
 		}
 
 		$pageId = $this->xmlHelper->getIDNodeValue( $objectNode );
